+++ src/message/index.tsx
@@ -1,16 +1,42 @@
 import styled from 'styled-components'
-import { color, space, font } from '_utils/branding'
+import { color, space, font, radius } from '_utils/branding'
 
 import Message from './Message'
 
 const StyledMessage = styled(Message)`
-  & {
+  & blockquote {
     margin: 0;
+    display: inline-block;
+    padding-bottom: ${space.s}
   }
+  
+  & {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  & .kirk-message-content {
+    word-break: break-word;
+  }
+  
+  & .kirk-message-content,
+  & .kirk-message-annotation {
+    display: flex;
+    margin-left: ${space.xxl};
+    margin-right: ${space.xxl};
+  }
+  
+  &.kirk-active .kirk-message-content,
+  &.kirk-active .kirk-message-annotation {
+    margin-left: 96px;
+    margin-right: 0;
+    justify-content: flex-end;
+  }
+  
   & .kirk-label {
     position: relative;
     padding: ${space.l};
-    border-radius: 20px;
+    border-radius: ${radius.l};
     border-top-left-radius: 0;
     background: ${color.lightBackground};
     color: ${color.primaryText};
@@ -21,29 +47,19 @@ const StyledMessage = styled(Message)`
   & .kirk-label p {
     margin: 0;
   }
-
+  
   &.kirk-active .kirk-label {
     background: ${color.primaryText};
     color: ${color.white};
-    border-top-left-radius: 20px;
+    border-top-left-radius: ${radius.l};
     border-bottom-right-radius: 0;
   }
-
-  & cite {
-    display: block;
-    margin-bottom: ${space.m};
-    color: ${color.secondaryText};
-    font-style: normal;
+  
+  & .kirk-message-annotation {
+    padding-left: ${space.l};
   }
-
-  &.active cite {
-    text-align: right;
-  }
-
-  & .kirk-date {
-    font-size: ${font.s.size};
-    line-height: ${font.s.lineHeight};
-    color: ${color.secondaryText};
+  &.kirk-active .kirk-message-annotation {
+    padding-right: ${space.l};
   }
 `