+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - **[NEW]** `focus` prop on `TimePicker`
 
+# v14.0.0 (08/11/2019)
+
+- **[FIX]** Restore outline on `Button`, `Dropdown`, `HamburgerButton` and `ToggleButton`
+- **[BREAKING CHANGE]** - Overhaul on `Message` component: Correct layout and support for message annotation.
+
 # v13.0.3 (06/11/2019)
 
 - **[FIX]** Fix scrolling visual for `Textarea`.